test(plugin-express): improvements and split files
