+++ packages/gasket-plugin-express/package.json
@@ -85,6 +85,14 @@
         "rules": {
           "jsdoc/*": "off"
         }
+      },
+      {
+        "files": [
+          "test/**/*.js"
+        ],
+        "rules": {
+          "jsdoc/require-jsdoc": "off"
+        }
       }
     ]
   },
